pyproject.toml
[project]
name = "image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
